+++ tests/Traits/PdfSectorTraitTest.php
@@ -24,7 +24,23 @@ class PdfSectorTraitTest extends TestCase
         self::assertSame(0, $document->getPage());
     }
 
-    public function testNegativeAngle(): void
+    public function testNegativeBothAngles(): void
+    {
+        $document = new PdfDocumentSector();
+        $document->addPage();
+        $document->sector(100, 100, 50, -10, -90);
+        self::assertSame(1, $document->getPage());
+    }
+
+    public function testNegativeEndAngle(): void
+    {
+        $document = new PdfDocumentSector();
+        $document->addPage();
+        $document->sector(100, 100, 50, 10, -90);
+        self::assertSame(1, $document->getPage());
+    }
+
+    public function testNegativeStartAngle(): void
     {
         $document = new PdfDocumentSector();
         $document->addPage();